+++ examples/api/api_collections.js
@@ -1,29 +1,31 @@
 import React from 'react';
-import Api from '../../src/index';
+import Api from '../../src/api';
 
-const schema = require( '../schema.json' );
-
-function create( sources ) {
-  sources
-    .$create({ title: 'foo' })
+function create( users ) {
+  users
+    .$create({ name: 'foo' }, { refetch: false })
     .then( data => console.log( 'Successfully created', data ) );
 }
 
-function Sources({ sources }) {
+function Users({ users }) {
   return (
     <div>
-      { sources.$list().$isPending ?
+      { users.$list().$isPending ?
         <span>Loading...</span>
       :
         <span>
-          {sources.$list().map( source =>
-            <div key={ source.get( '_id' ) }>
-              <span>{ source.get( 'title' ) }</span>
-              <button onClick={() => source.$delete()}>Delete</button>
+          {users.$list().map(( user, idx ) =>
+            <div key={ user.get( 'id' ) }>
+              <span>{ user.get( 'name' ) }</span>
+              { idx < 10 &&
+                <button onClick={() => user.$delete()}>Delete</button>
+              }
             </div>
           )}
-          <button onClick={() => create( sources )}>New foo</button>
-          <button onClick={() => sources.$clear()}>Clear cache</button>
+          {users.$list().size === 10 &&
+            <button onClick={() => create( users )}>New foo</button>
+          }
+          <button onClick={() => users.$list().$clear()}>Clear cache</button>
         </span>
       }
     </div>
@@ -32,8 +34,8 @@ function Sources({ sources }) {
 
 export default function ApiCollectionsExample() {
   return (
-    <Api url="http://localhost:3000/hub/api" schema={ schema }>
-      <Sources />
+    <Api url="https://jsonplaceholder.typicode.com" users="/users">
+      <Users />
     </Api>
   );
 }