+++ src/lib/schema.js
@@ -41,17 +41,18 @@ export function getIdFromModel( model ) {
     : model._id || model.id || model.ID;
 }
 
-export function isSchema( maybeSchema ) {
+export const isSchema = props => key => {
+  const maybeSchema = props[ key ];
   return (
     maybeSchema && ( 
       typeof maybeSchema === 'object' && 
-      ( !!maybeSchema.$schema || !!maybeSchema.schema )
+      ( !!maybeSchema.$schema || !!maybeSchema.schema || !!maybeSchema.routes )
     ) || (
       typeof maybeSchema === 'string' &&
       maybeSchema.indexOf( '/' ) === 0
     )
   );
-}
+};
 
 export function getSchema( maybeSchema ) {
   return (
@@ -66,5 +67,4 @@ export function getSchema( maybeSchema ) {
       ? { title: maybeSchema.slice( 1 ) }
       : null
   );
-
 }