+++ src/store.js
@@ -17,12 +17,16 @@ export function managesState() {
       );
     },
     setState( path, val ) {
-      const immutableVal = fromJS( val );
-      _state = (
-        Array.isArray( path ) ?
-          _state.setIn( path, immutableVal ) :
-        _state.set( path, immutableVal )
-      );
+      if ( typeof val !== 'undefined' ) {
+        const immutableVal = fromJS( val );
+        _state = (
+          Array.isArray( path ) ?
+            _state.setIn( path, immutableVal ) :
+          _state.set( path, immutableVal )
+        );
+      } else {
+        _state = path;
+      }
       return _state;
     },
     resetState( val, isUndo = true ) {
@@ -52,7 +56,7 @@ export function offersSubscription() {
   };
 }
 
-export function canUndo({ apply = () => {} }) {
+export function canUndo({ apply = () => {} } = {}) {
 
   const log = ( ...args ) => logger( '\u23f1 timetravel: ', ...args );
 
@@ -64,9 +68,9 @@ export function canUndo({ apply = () => {} }) {
       const current = undo.get( idx );
       const next = undo.get( idxNext );
       const changes = diff( current, next ).toJS();
-      if ( !ignore || !changes.every( item => item.path.indexOf( `/${ignore}` ) === 0 ) ) {
+      idx = idxNext;
+      if ( !ignore || !changes.every( item => item.path.indexOf( ` ‣ ${ignore}` ) === 0 ) ) {
         log( `diff:`, changes );
-        idx = idxNext;
         apply( next );
         return true;
       } else {
@@ -74,12 +78,12 @@ export function canUndo({ apply = () => {} }) {
         return false;
       }
     },
-    prev({ ignore }) {
+    prev({ ignore } = {}) {
       const idxNext = idx - 1;
       if ( idxNext >= -(undo.size) ) {
         const display = idx * -1;
         log(`stepping back to ${display} state(s) ago` );
-        if ( !methods.apply( idxNext ) ) {
+        if ( !methods.apply( idxNext, ignore ) ) {
           methods.prev({ ignore });
         }
       } else {
@@ -88,11 +92,11 @@ export function canUndo({ apply = () => {} }) {
         log(`there are no earlier states than this one`);
       }
     },
-    next({ ignore }) {
+    next({ ignore } = {}) {
       if ( idx < -1 ) {
         const idxNext = idx + 1;
         log(`stepping forward to ${idx === -2 ? 'current state' : (idx + 2) * -1 + ' state(s) ago' }`);
-        if ( !methods.apply( idxNext ) ) {
+        if ( !methods.apply( idxNext, ignore ) ) {
           methods.next({ ignore });
         }
       } else {
@@ -108,6 +112,7 @@ export function canUndo({ apply = () => {} }) {
     },
     save( state ) {
       undo = undo.push( state );
+      return undo;
     },
     isDirty() {
       return idx !== -1;
@@ -117,49 +122,59 @@ export function canUndo({ apply = () => {} }) {
   return methods;
 }
 
-const { invoke, subscribe, ...subscriptionMethods } = offersSubscription();
-const { setState, resetState, ...stateMethods } = managesState();
+export function initStore({ 
+  subscriptionInit = offersSubscription,
+  stateInit        = managesState,
+  undoInit         = canUndo
+} = {}) {
 
-// TODO: clarify options
-const undo = canUndo({ apply: ( state, options = {} ) => invoke( resetState( state ), !options.reset ) } );
+  const { invoke, subscribe, ...subscriptionMethods } = subscriptionInit();
+  const { setState, resetState, ...stateMethods } = stateInit();
 
-const setStateEmoji = '\uD83C\uDFDB';
+  // TODO: clarify options
+  const undo = undoInit({ apply: ( state, options = {} ) => invoke( resetState( state ), !options.reset ) } );
 
-const store = {
-  ...subscriptionMethods,
-  ...stateMethods,
-  setState( ...args ) {
-    if ( undo.isDirty() ) undo.reset();
-    const statePrev = stateMethods.getState();
-    const stateNext = setState( ...args );
-    undo.save( stateNext );
-    logger( `${setStateEmoji} setting state: `, diff( statePrev, stateNext ).toJS() );
-    // TODO: bump into next event loop to avoid possible collisions?
-    invoke( stateNext );
-  },
-  setStateQuiet( ...args ) {
-    const statePrev = stateMethods.getState();
-    const stateNext = setState( ...args );
-    logger( `%c${setStateEmoji} setting state quiet (no re-rendering):`, `color: #999`, diff( statePrev, stateNext ).toJS() );
-  },
-  subscribeAll: subscribe,
-  subscribeTo( path, callback, initialState ) {
-    let cache = null;
-    const onChange = state => {
-      const nextState = state.getIn( [].concat( path ) );
-      if ( nextState !== cache ) {
-        callback( nextState );
-        cache = nextState;
-      }
-    };
-    subscribe( onChange );
-    if ( initialState ) setState( path, initialState );
-    return onChange;
-  },
-  nextState: undo.next,
-  prevState: undo.prev,
-  resetState: undo.reset
-};
+  const setStateEmoji = '\uD83C\uDFDB';
 
-export default store;
+  const store = {
+    ...subscriptionMethods,
+    ...stateMethods,
+    setState( ...args ) {
+      if ( undo.isDirty() ) undo.reset();
+      const statePrev = stateMethods.getState();
+      const stateNext = setState( ...args );
+      undo.save( stateNext );
+      logger( `${setStateEmoji} setting state: `, diff( statePrev, stateNext ).toJS() );
+      // TODO: bump into next event loop to avoid possible collisions?
+      invoke( stateNext );
+    },
+    setStateQuiet( ...args ) {
+      const statePrev = stateMethods.getState();
+      const stateNext = setState( ...args );
+      logger( `%c${setStateEmoji} setting state quiet (no re-rendering):`, `color: #999`, diff( statePrev, stateNext ).toJS() );
+      return stateNext;
+    },
+    subscribeAll: subscribe,
+    subscribeTo( path, callback, initialState ) {
+      let cache = null;
+      const onChange = state => {
+        const nextState = state.getIn( [].concat( path ) );
+        if ( nextState !== cache ) {
+          callback( nextState );
+          cache = nextState;
+        }
+      };
+      subscribe( onChange );
+      if ( initialState ) setState( path, initialState );
+      return onChange;
+    },
+    clearState: () => store.setState( Map({}) ),
+    nextState:  undo.next,
+    prevState:  undo.prev,
+    resetState: undo.reset
+  };
 
+  return store;
+}
+
+export default initStore();